requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
langchain==0.0.350
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
openai==1.3.5
tiktoken==0.5.1 